Changed file name from mapGen to worldGen. Started the file opening and setup process for worldGen.cc
