--- mapGen.cc
@@ -1,25 +1,0 @@
-
-/****************************************************************************************************/
-// Global includes
-
-#include <random>
-
-/****************************************************************************************************/
-// Local includes
-
-
-/****************************************************************************************************/
-// Using statements
-
-using std::vector;
-
-typedef uchar unsigned char
-/****************************************************************************************************/
-// Main
-
-int main(int args, char* argv[])
-{
-    vector<vector<uchar>> world;
-    
-    return EXIT_SUCCESS;
-}
